% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munging_functions.R
\name{booleanize}
\alias{booleanize}
\title{Reduce the numeric values of a data.frame to boolean values}
\usage{
booleanize(matrix, present = TRUE, absent = FALSE)
}
\arguments{
\item{matrix}{data.frame with numeric values}

\item{present}{replacement values for cells with numeric value >0. default: TRUE}

\item{absent}{replacement values for cells with numeric value 0. default: FALSE}
}
\value{
data.frame with present-absent values
}
\description{
\code{booleanize} returns an other version of the input data.frame with
simple, definable present-absent information instead of numeric values.
}
\examples{
testmatrix <- data.frame(c1 = c(0,2,0,8), c2 = c(5,6,7,0), c3 = c(5,6,7,0))

booleanize(testmatrix)
booleanize(testmatrix, 5, 10)

cakedata <- booleanize(matrix = testmatrix, present = "cake", absent = "no cake")

}

