% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munging_functions.R
\name{check_num_nonzero}
\alias{check_num_nonzero}
\title{Checks if the number of non zero elements in a vector is at least equal to a given parameter.}
\usage{
check_num_nonzero(x, minnumber)
}
\arguments{
\item{x}{The vector to be tested.}

\item{minnumber}{A number of minimum non zero elements.}
}
\value{
The \code{true} if \code{x} has at least \code{minnumber}  non zero elements,
else \code{false}.
}
\description{
Checks if the number of non zero elements in a vector is at least equal to a given parameter.
}
\examples{
check_num_nonzero(c(0,0,0,0,0), 1)
check_num_nonzero(c(0,0,0,0,1), 1)
check_num_nonzero(c(2:4,0,1), 1)

}

