% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_utility_functions.R
\name{delempty}
\alias{delempty}
\title{Delete "empty" rows and columns of a data.frame}
\usage{
delempty(matrix)
}
\arguments{
\item{matrix}{data.frame with numeric values}
}
\value{
data.frame without empty rows and columns (matrix[-z,-n])
}
\description{
A row or a column of a data.frame with numeric values is considered empty, if all
its values are 0. \code{delempty} deletes all empty rows and columns of an input
data.frame.
}
\examples{
testmatrix <- data.frame(c1 = c(0,3,8,2), c2 = c(0,6,7,8), c3 = c(0,0,0,0))

delempty(testmatrix)

withoutempty <- delempty(matrix = testmatrix)

}
