% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delrc.R
\name{delrc}
\alias{delrc}
\title{Delete rows and columns of a data.frame by amount of values > 0}
\usage{
delrc(matrix, climit = 0, rlimit = 0)
}
\arguments{
\item{matrix}{data.frame with numeric values}

\item{climit}{numeric value: How many values > 0 have to be present to consider a 
column sufficiently linked to perform further analysis? Every column with less values > 0 
will be removed. 

default = 0 (no column will be deleted)}

\item{rlimit}{numeric value: How many values > 0 have to be present to consider a 
row sufficiently linked to perform further analysis? Every row with less values > 0 
will be removed. 

default = 0 (no row will be deleted)}
}
\value{
data.frame without removed rows/colums. 

If no values or just one row are left in the data.frame after application of the selection 
criteria, the output is FALSE.
}
\description{
A data.frame with numeric values often contains rows and columns with an insufficient 
amount of values > 0 for a certain task. For example correspondence analysis or 
bivariate correlation analysis requires a minimum amount of usable values. 
In an archaeological context this could apply for example, if certain find categories
are particularly rare in a burial site context.
\code{delrc} allows to remove rows and columns, that don't fulfill the requirements.
}
\examples{
testmatrix <- data.frame(
c1 = c(0,3,8,2), 
c2 = c(0,6,7,0), 
c3 = c(0,0,0,0),
c4 = c(0,3,8,2), 
c5 = c(0,6,7,0), 
c6 = c(1,0,0,1)
)

# The following code removes every column with less than 3 values > 0.
# That will remove the columns c2, c3, c5 and c6.
# Further, every row with less than 2 values gets removed. 
# That will delete row 1.
delrc(testmatrix, climit = 3, rlimit = 2)

testresult <- delrc(testmatrix, 3, 2)
rowSums(booleanize(testresult))
colSums(booleanize(testresult))

}

