% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fillhexa}
\alias{fillhexa}
\title{Fills hexahedrons with a systematic point raster}
\usage{
fillhexa(hex, res)
}
\arguments{
\item{hex}{matrix}

\item{res}{double}
}
\value{
final matrix
}
\description{
test
}
\examples{
hexatestdf <- data.frame(
  x = c(0,1,0,4,5,5,5,5),
  y = c(1,1,4,4,1,1,4,4),
  z = c(4,8,4,9,4,8,4,6)
)

hexatest <- as.matrix(hexatestdf)

cx = fillhexa(hexatest, 0.1)

library(rgl)
plot3d(
 cx[,1], cx[,2], cx[,3],
 type = "p",
 xlab = "x", ylab = "y", zlab = "z"
)

}

