% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{igraphmask}
\alias{igraphmask}
\title{Mask function for igraph::plot.igraph() to visualize a relation table as a graph}
\usage{
igraphmask(reltable, mypath, w = 3000, h = 3000,
  colorvector = c("red"))
}
\arguments{
\item{reltable}{list of bivariate relations (as produced by quantAAR::reltable())}

\item{mypath}{file.path where the plot file should be stored

mypath <- file.path("~/path/to/my/directory/",paste("myfilename", ".png", sep = ""))}

\item{w}{width of the resulting graphic file in px

default = 3000}

\item{h}{width of the resulting graphic file in px

default = 3000}

\item{colorvector}{string vector with colour values to mark certain vertices

default = c("red")}
}
\value{
graphic file of a graph plot
}
\description{
\code{igraphmask} is an input mask for the \code{plot.igraph()} function of the
package igraph. \code{plot.igraph()} creates graph plots based on bivariate relations.
See \code{?igraph} for further info.
\code{igraphmask} allows to get a nice, basic graph plot as a png-file
that is directly saved to the file system. It doesn't allow to define specific
plot settings.
}
\details{
For graph layout the fruchterman.reingold.grid algorithm is used.
}
\examples{
testmatrixrand <- data.frame(
   matrix(base::sample(0:1,400,replace=TRUE), nrow=20, ncol=20)
)

testcorr <- corrmat(testmatrixrand, "lambda", chi2limit = 0.1, dim = 1)
testrel <- reltable(testcorr)

testpath <- file.path(".",paste("testfile", ".png", sep = ""))

igraphmask(testrel, testpath, w = 1000, h = 1000)

}
