% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_utility_functions.R
\name{itremove}
\alias{itremove}
\title{Removes iterative all rows and columns of a matrix or dataframe with less than a given number of non zero elements.}
\usage{
itremove(object, minnumber)
}
\arguments{
\item{object}{A matrix or dataframe.}

\item{minnumber}{A number of minimum non zero elements.}
}
\value{
A matrix or dataframe with all rows and columns removed that had less than the given number of non zero elements.
}
\description{
Removes iterative all rows and columns of a matrix or dataframe with less than a given number of non zero elements.
}
\examples{
a <- matrix(c(0,1,0,0,0,0,0,0,1,0,1,0,0,1,0,0),nrow=4)
a
itremove(a,1)

}
