% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{kriglist}
\alias{kriglist}
\title{Apply kriging for a list of data.frames}
\usage{
kriglist(plist, x = 1, y = 2, z = 3, lags = 6, model = "spherical")
}
\arguments{
\item{plist}{test}

\item{x}{test}

\item{y}{test}

\item{z}{test}

\item{lags}{test}

\item{model}{test}
}
\value{
list with data.frame maps
}
\description{
Test
}
\examples{
1 == 1

}

