% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage.R
\name{linkage}
\alias{linkage}
\title{Calculate linkage values of variables and objects of a numeric data.frame}
\usage{
linkage(matrix)
}
\arguments{
\item{matrix}{data.frame with numeric values}
}
\value{
table with linkage values
}
\description{
\code{linkage()} calculates a linkage value for every variable and every object
of an input data.frame and returns a table with this information. 
This linkage value allows predictions about whether a variable/object can be used
for multivariate analysis. If a variable/object is not well linked to the other entities, 
it will often appear as an outlier.
}
\details{
Structure of the resulting table: 

column 1:  linkage value (see code to understand how it's calculated)

column 2:  logarithm of linkage value (useful for interpretation)

column 3:  type (variable or object)
}
\examples{
testmatrixrand <- data.frame(
   matrix(base::sample(0:1,400,replace=T), nrow=20, ncol=20)
)
 
linkage(testmatrixrand)
 
link <- subset(linkage(testmatrixrand), type == "obj")  
barplot(link$linkage, names.arg = rownames(link))  
 
}

