% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{matuskovo}
\alias{matuskovo}
\alias{matuskovo_material}
\title{A burial site from the Unetice culture in western Slovakia}
\format{A tibble (data.frame) with 62 rows and 34 variables:
\itemize{
  \item grave_nr: number of the grave in Tocik 1979
  \item burial_depth_in_cm: numeric
  \item burial_length_in_cm: numeric
  \item burial_width_in_cm: numeric
  \item orientation: factor variable of grave orientation
  \item sex: factor variable of the anthropological determination
  \item posture: factor variable of the posture of the skeleton
  \item ceramics_general: count of ceramic items that could not be classified further
  \item pots: count
  \item bowls: count
  \item cups: count
  \item jugs: count
  \item pearls_bronze: count
  \item pearls_amber: count
  \item spiral_bead_bronze: count
  \item locket_shell: count
  \item temple_rings_big: count
  \item temple_rings_small: count
  \item ring_gold: count
  \item ring_bronze: count
  \item burl_ring: count
  \item arm_bracelet: count
  \item arm_spiral: count
  \item roll.top_pin: count
  \item eylet_pin: count
  \item dagger: count
  \item misc_shells: count
  \item misc_stones: count
  \item misc_jaspis: count
  \item misc_bones: count
  \item misc_animal_teeth: count
  \item misc_bronze: count
  \item misc_clay_tuyere: count
  \item notes: character variable of additional notes
}}
\source{
{Tocik, Anton (1979) Vycapy-Opatovce : a dalsie pohrebiska z
 starsej doby bronzovej na Juhozapadnom Slovensku = Vycapy-Oopatovce :
 und weitere altbronzezeitliche Graberfelder in der Sudwestslowakei. Nitra.}
}
\description{
\code{matuskovo} is a tibble containing graves (objects) and grave attributes
(variables) of a burial site from the Unetice culture in western Slovakia.
\code{matuskovo_material} is a matrix subset with only the artefact count variables.
}
