% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munging_functions.R
\name{newcorrtable}
\alias{newcorrtable}
\title{Create an empty correlation matrix of a given data.frame}
\usage{
newcorrtable(matrix, dim = 1)
}
\arguments{
\item{matrix}{data.frame}

\item{dim}{switch to define, whether the new correlation matrix should be created
for columns or rows.

1 (default): table is created for column (variables) relations.

2: table is created for row (objects) relations.}
}
\value{
empty correlation matrix data.frame
}
\description{
\code{newcorrtable} returns an empty correlation matrix of a data.frames columns
or rows. Empty means: filled with the numeric value "0".
This correlation matrix data.frame can be used to store the results of functions which
calculate correlation values for all bivariate relations.
}
\examples{
testmatrix <- data.frame(c1 = c(5,2,3,8), c2 = c(5,6,7,0), c3 = c(5,6,7,9))

# correlation table is created for the columns of the input data.frame testmatrix
newcorrtable(testmatrix)
newcorrtable(testmatrix, 1)
corrtabcolumns <- newcorrtable(matrix = testmatrix, dim = 1)

# correlation table is created for the rows of the input data.frame testmatrix
newcorrtable(testmatrix, 2)
corrtabrows <- newcorrtable(matrix = testmatrix, dim = 2)

}

