% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posdec}
\alias{posdec}
\title{Test}
\usage{
posdec(crlist, maplist)
}
\arguments{
\item{crlist}{TODO}

\item{maplist}{TODO}
}
\description{
Testdescription
}
\examples{
df1 <- data.frame(
  x = rnorm(50),
  y = rnorm(50),
  z = rnorm(50) - 5
)

df2 <- data.frame(
  x = rnorm(50),
  y = rnorm(50),
  z = rnorm(50) + 5
)

lpoints <- list(df1, df2)

maps <- kriglist(lpoints, lags = 3, model = "spherical")

finds <- data.frame(
  x = c(0, 1, 0.5, 0.7),
  y = c(0.5, 0, 1, 0.7),
  z = c(-10, 10, 0, 2)
)

posdec(finds, maps)

}

