% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posdeclist}
\alias{posdeclist}
\title{Test}
\usage{
posdeclist(crlist, maplist)
}
\arguments{
\item{crlist}{TODO}

\item{maplist}{TODO}
}
\description{
Testdescription
}
\examples{
df1 <- data.frame(
  x = rnorm(50),
  y = rnorm(50),
  z = rnorm(50) - 5
)

df2 <- data.frame(
  x = rnorm(50),
  y = rnorm(50),
  z = rnorm(50) + 5
)

lpoints <- list(df1, df2)

maps <- kriglist(lpoints, lags = 3, model = "spherical")

hexadf1 <- data.frame(
  x = c(0, 1, 0, 4, 5, 5, 5, 5),
  y = c(1, 1, 4, 4, 1, 1, 4, 4),
  z = c(1, 5, 1, 6, 1, 5, 1, 3)
)

hexadf2 <- data.frame(
  x = c(0, 1, 0, 4, 5, 5, 5, 5),
  y = c(1, 1, 4, 4, 1, 1, 4, 4),
  z = c(-1, -5, -1, -6, -1, -5, -1, -3)
)

hexa1 <- as.matrix(hexadf1)
hexa2 <- as.matrix(hexadf2)

cx1 <- fillhexa(hexa1, 0.1)
cx2 <- fillhexa(hexa2, 0.1)

cubelist <- list(cx1, cx2)

posdec(cubelist, maps)

}

