% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_munging_functions.R
\name{presencecount}
\alias{presencecount}
\title{Count appearence of variables and objects in a data.frame with numeric values}
\usage{
presencecount(matrix, dim = 1)
}
\arguments{
\item{matrix}{data.frame with numeric values}

\item{dim}{switch to define if the appearences in columns (variables) or rows (objects)
should be counted.

1: column (variables) appearences are counted

2: row (objects) appearences are counted}
}
\value{
sorted data.frame with amount of appearences for objacts or variables
}
\description{
A row or a column of a data.frame with numeric values is considered empty, if all
its values are 0. If objects or variables are present, their rows and columns contain
values != 0. To count the appearence of objects or variables, \code{presencecount} counts
the amount of values != 0 within the rows or columns and writes them into a data.frame.
The result is sorted by the amount of appearences.
}
\examples{
testmatrix <- data.frame(c1 = c(0,3,8,2), c2 = c(0,6,7,8), c3 = c(0,0,0,0))
rownames(testmatrix) <- c("r1","r2","r3","r4")

countvar <- presencecount(testmatrix, 1)
countobj <- presencecount(testmatrix, 2)

}
