% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_curve.R
\name{principal_curve}
\alias{principal_curve}
\alias{principal_curve.princurve_principal_curve}
\title{Principal Curves}
\usage{
principal_curve.princurve_principal_curve(..., raw_output = TRUE)
}
\arguments{
\item{...}{Input arguments of the relevant wrapped functions.}

\item{raw_output}{Logical. Should the raw output of the wrapped functions be stored as
an additional output attribute "raw"? Default: TRUE.}
}
\value{
A tibble with the modified input point coordinates after the projection onto the curve.
Additional values are stored in object attributes. See \code{attributes(result)$raw}.

...: Additional variables as provided by the wrapped functions.

X1...XX: Numeric. Resulting coordinates.
}
\description{
Principal Curve fitting function wrappers that give the result in a tidy data.frame.
}
\examples{
principal_curve.princurve_principal_curve(
  as.matrix(ca.ca_ca(matuskovo_material)[, c("CA1", "CA2", "CA3")])
)

}
