% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriation.R
\name{seriation.seriation_seriate}
\alias{seriation.seriation_seriate}
\alias{seriation2widedf}
\alias{seriation.tabula_seriate}
\title{Seriation}
\usage{
seriation.seriation_seriate(...)

seriation2widedf(x)

seriation.tabula_seriate(...)
}
\arguments{
\item{...}{Input arguments of \code{seriation::seriate}.}

\item{x}{Data.frame. Output of the seriation wrapper functions.}
}
\value{
A tibble with the seriation result in a long format.

row: Character. Names of rows.

row_order: Integer. Order of rows.

col: Character. Names of columns.

col_order: Integer. Order of columns.

value: Numeric. Value in input matrix for respective row and col.
}
\description{
Seriation function wrappers that give the result in a long tidy data.frame.
The result table can be transformed to a wide format with \code{seriation2widedf}.
}
\examples{
## seriate matrix
data("iris")
x <- iris[-5]

## to make the variables comparable, we scale the data
x <- scale(x, center = FALSE)
x <- as.data.frame(x)

## try some methods
order <- quantAAR::seriation.seriation_seriate(x, method = "PCA")

#x_matrix <- as(x, "IncidenceMatrix")
#order <- quantAAR::seriation.tabula_seriate(x_matrix, method = "correspondance")

## transform back to wide format
seriation2widedf(order)

}
