% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{spatialwide}
\alias{spatialwide}
\title{Transform a long surface describing df to a wide surface describing df}
\usage{
spatialwide(x, y, z)
}
\arguments{
\item{x}{test}

\item{y}{test}

\item{z}{test}
}
\value{
widedf
}
\description{
Test
}
\examples{
1 == 1

}

