% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_utility_functions.R
\name{tidyca}
\alias{tidyca}
\title{Mask function for ca:ca() to get the ca results in a tidy data.frame}
\usage{
tidyca(x, ...)
}
\arguments{
\item{x}{data.frame with numeric values}

\item{...}{further options of \code{ca::ca}}
}
\value{
data.frame with ca coordinates of variables and objects of the input data.frame
}
\description{
\code{tidyca()} delivers the coordinates of objects and variables of a correspondence
analyse as a tidy data.frame. The CA is calculated by \code{ca::ca()}.
See \code{?ca} for further information.
}
\examples{
testmatrixrand <- data.frame(
   matrix(base::sample(0:1,400,replace=TRUE), nrow=20, ncol=20)
)
rownames(testmatrixrand) <- paste("row", seq(1:nrow(testmatrixrand)))

tidyca(testmatrixrand, supc = c(1,2,3), supr = c(15,16))

}
