% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriation_utility_functions.R
\name{tidyseriation}
\alias{tidyseriation}
\alias{seriation2widedf}
\title{tidyseriation}
\usage{
tidyseriation(...)

seriation2widedf(x)
}
\arguments{
\item{...}{Input arguments of \code{seriation::seriate}.}

\item{x}{Data.frame. Output of tidyseriation.}
}
\value{
A tibble with the seriation result in a long format.

row: Character. Names of rows.

row_order: Integer. Order of rows.

col: Character. Names of columns.

col_order: Integer. Order of columns.

value: Numeric. Value in input matrix for respective row and col.
}
\description{
Mask function for seriation::seriate() to get the seriation results in a tidy data.frame.
The result table can be transformed back to a wide format with \code{seriation2widedf}.
}
\details{
\code{tidyseriation()} reorders the input table with seriation and gives back the result in a long tidy data.frame.
The seriation is calculated by \code{seriation::seriate()}.
See \code{?seriate} for further information.
}
\examples{
testmatrixrand <- data.frame(
   matrix(base::sample(0:1,400,replace=TRUE), nrow=20, ncol=20)
)
rownames(testmatrixrand) <- paste0("row", seq(1:nrow(testmatrixrand)))
colnames(testmatrixrand) <- paste0("col", seq(1:ncol(testmatrixrand)))

tidyseriation(testmatrixrand, method = "PCA")

# transform back to wide format
seriation2widedf(tidyseriation(testmatrixrand, method = "PCA"))

}
