% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{ca.ca_ca}
\alias{ca.ca_ca}
\alias{get_dimension_label}
\alias{ca.vegan_cca}
\title{Correspondence Analysis}
\usage{
ca.ca_ca(...)

get_dimension_label(x, dim)

ca.vegan_cca(...)
}
\arguments{
\item{...}{Input arguments of the relevant wrapped functions.}

\item{x}{Output of \code{ca.ca_ca}.}

\item{dim}{Index of dimension.}
}
\value{
A tibble with the ca results for variables (columns) and objects (rows).
Additional values are stored in object attributes.

name: Character. Names of rows and columns.

type: Character. Type of entry ("row" or "col").

...: Additional variables as provided by the wrapped functions.

x1...xX: Numeric. Standard coordinates of this entry in all available dimensions.
}
\description{
Transforms the output coordinates of objects and variables of a correspondence
analysis to a tidy data.frame.
}
\examples{
haireye <- margin.table(datasets::HairEyeColor, 1:2)
quantAAR::ca.ca_ca(haireye)
quantAAR::ca.vegan_cca(haireye)

}
