% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.ca_ca.R
\name{ca.ca_ca}
\alias{ca.ca_ca}
\alias{get_dimension_label}
\title{Correspondence Analysis}
\usage{
ca.ca_ca(...)

get_dimension_label(x, dim)
}
\arguments{
\item{...}{Input arguments of \code{ca::ca}.}

\item{x}{Output of \code{ca.ca_ca}.}

\item{dim}{Index of dimension.}
}
\value{
A tibble with the ca results for variables (columns) and objects (rows).
The singular values and simplified dimension weights are stored as attributes to
the tibble and can be accessed via \code{attr(res, "singluar_values")} and
\code{attr(res, "simplified_dimension_weights")}.

name: Character. Names of rows and columns.

type: Character. Type of entry ("row" or "col").

sup: Boolean. Was this entry treated as a supplementary point?

mass: Mumeric. Row and column masses.

dist: Mumeric. Row and column chi-square distances to centroid.

mass: Mumeric. Row and column inertias.

Dim1...DimX: Mumeric. Standard coordinates of this entry in all available dimensions.
}
\description{
Transforms the output coordinates of objects and variables of a correspondence
analysis to a tidy data.frame. The CA is calculated by \code{ca::ca()}.
}
\examples{
haireye <- margin.table(datasets::HairEyeColor, 1:2)
quantAAR::ca.ca_ca(haireye)

}
