% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.stats_prcomp}
\alias{pca.stats_prcomp}
\title{Principal Components Analysis}
\usage{
pca.stats_prcomp(...)
}
\arguments{
\item{...}{Input arguments of \code{stats::prcomp}.}
}
\value{
A tibble with the PCA results for variables (columns) and objects (rows).
The standard deviations of the principal components and the centering and scaling used
are stored as attributes to the tibble and can be accessed via
\code{attr(res, "standard_deviations")},
\code{attr(res, "center")} and \code{attr(res, "scale")}.

name: Character. Names of rows and columns.

type: Character. Type of entry ("row" or "col").

PC1...PCX: Numeric. Resulting coordinates in all principal component dimensions.
}
\description{
Transforms the output coordinates of objects and variables of a principal components
analysis to a tidy data.frame. The PCA is calculated by \code{stats::prcomp()}.
}
\examples{
quantAAR::pca.stats_prcomp(datasets::USArrests)

}
