% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.stats_prcomp}
\alias{pca.stats_prcomp}
\title{Principal Components Analysis}
\usage{
pca.stats_prcomp(...)
}
\arguments{
\item{...}{Input arguments of the relevant wrapped functions.}
}
\value{
A tibble with the PCA results for variables (columns) and objects (rows).
Additional values are stored in object attributes. See \code{attributes(result)}.

name: Character. Names of rows and columns.

type: Character. Type of entry ("row" or "col").

...: Additional variables as provided by the wrapped functions.

x1...xX: Numeric. Resulting coordinates in all principal component dimensions.
}
\description{
Principal Components Analysis function wrappers that give the result in a tidy data.frame.
}
\examples{
# establish an other distance measure in the data that allows application of PCA
matuskovo_distance <- vegan::decostand(matuskovo_material, "norm")

pca.stats_prcomp(matuskovo_distance)

}
