% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\alias{pca.stats_prcomp}
\title{Principal Components Analysis}
\usage{
pca.stats_prcomp(..., raw_output = TRUE)
}
\arguments{
\item{...}{Input arguments of the relevant wrapped functions.}

\item{raw_output}{Logical. Should the raw output of the wrapped functions be stored as
an additional output attribute "raw"? Default: TRUE.}
}
\value{
A tibble with the PCA results for variables (columns) and objects (rows).
Additional values are stored in object attributes. See \code{attributes(result)$raw}.

name: Character. Names of rows and columns.

type: Character. Type of entry ("row" or "col").

...: Additional variables as provided by the wrapped functions.

x1...xX: Numeric. Resulting coordinates in all principal component dimensions.
}
\description{
Principal Components Analysis function wrappers that give the result in a tidy data.frame.
}
\examples{
# establish an other distance measure in the data that allows application of PCA
matuskovo_distance <- vegan::decostand(matuskovo_material, "norm")

pca.stats_prcomp(matuskovo_distance)

}
