% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posdec}
\alias{posdec}
\title{position decision in relation to a set of stacked surfaces}
\usage{
posdec(crlist, maplist)
}
\arguments{
\item{crlist}{data.frame with the spatial coordinates of the points of interest}

\item{maplist}{list of data.frames which contain the points that make up the surfaces}
}
\value{
data.frame with the spatial coordinates of the points of interest and the respective position
information
}
\description{
\code{posdec} has the purpose to make a decision about the position of individual points in relation
to a set of stacked surfaces. The decision is made by comparing the mean z-axis value of the four
horizontally closest points of a surface to the z-axis value of the point in question.
}
\examples{
df1 <- data.frame(
  x = rnorm(50),
  y = rnorm(50),
  z = rnorm(50) - 5
)

df2 <- data.frame(
  x = rnorm(50),
  y = rnorm(50),
  z = rnorm(50) + 5
)

lpoints <- list(df1, df2)

maps <- kriglist(lpoints, lags = 3, model = "spherical")

finds <- data.frame(
  x = c(0, 1, 0.5, 0.7),
  y = c(0.5, 0, 1, 0.7),
  z = c(-10, 10, 0, 2)
)

posdec(finds, maps)

}

