% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posdeclist}
\alias{posdeclist}
\title{position decision in relation to a set of stacked surfaces (for lists of data.frames)}
\usage{
posdeclist(crlist, maplist)
}
\arguments{
\item{crlist}{list of data.frames with the spatial coordinates of the points of interest}

\item{maplist}{list of data.frames which contain the points that make up the surfaces}
}
\value{
list of data.frames with the spatial coordinates of the points of interest and the respective
position information
}
\description{
\code{posdeclist} works as \code{posdec} but not just for a single data.frame but for a list of
data.frames
}
\examples{
df1 <- data.frame(
  x = rnorm(50),
  y = rnorm(50),
  z = rnorm(50) - 5
)

df2 <- data.frame(
  x = rnorm(50),
  y = rnorm(50),
  z = rnorm(50) + 5
)

lpoints <- list(df1, df2)

maps <- kriglist(lpoints, lags = 3, model = "spherical")

hexadf1 <- data.frame(
  x = c(0, 1, 0, 4, 5, 5, 5, 5),
  y = c(1, 1, 4, 4, 1, 1, 4, 4),
  z = c(1, 5, 1, 6, 1, 5, 1, 3)
)

hexadf2 <- data.frame(
  x = c(0, 1, 0, 4, 5, 5, 5, 5),
  y = c(1, 1, 4, 4, 1, 1, 4, 4),
  z = c(-1, -5, -1, -6, -1, -5, -1, -3)
)

cx1 <- fillhexa(hexadf1, 0.1)
cx2 <- fillhexa(hexadf2, 0.1)

cubelist <- list(cx1, cx2)

posdeclist(cubelist, maps)

}

