% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriation.R
\name{seriation.seriation_seriate}
\alias{seriation.seriation_seriate}
\alias{seriation2widedf}
\title{Seriation}
\usage{
seriation.seriation_seriate(...)

seriation2widedf(x)
}
\arguments{
\item{...}{Input arguments of \code{seriation::seriate}.}

\item{x}{Data.frame. Output of tidyseriation.}
}
\value{
A tibble with the seriation result in a long format.

row: Character. Names of rows.

row_order: Integer. Order of rows.

col: Character. Names of columns.

col_order: Integer. Order of columns.

value: Numeric. Value in input matrix for respective row and col.
}
\description{
Mask function for seriation::seriate() to get the seriation results in a tidy data.frame.
The seriation is calculated by \code{seriation::seriate()}.
The result table can be transformed back to a wide format with \code{seriation2widedf}.
}
\examples{
## seriate matrix
data("iris")
x <- iris[-5]

## to make the variables comparable, we scale the data
x <- scale(x, center = FALSE)
x <- as.data.frame(x)

## try some methods
order <- quantAAR::seriation.seriation_seriate(x, method = "PCA")

## transform back to wide format
seriation2widedf(order)

}
