% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriation.R
\name{seriation.seriation_seriate}
\alias{seriation.seriation_seriate}
\alias{spread_seriation}
\alias{seriation.tabula_seriate}
\title{Seriation}
\usage{
seriation.seriation_seriate(..., raw_output = TRUE)

spread_seriation(x)

seriation.tabula_seriate(..., raw_output = TRUE)
}
\arguments{
\item{...}{Input arguments of \code{seriation::seriate}.}

\item{raw_output}{Logical. Should the raw output of the wrapped functions be stored as
an additional output attribute "raw"? Default: TRUE.}

\item{x}{Data.frame. Output of the seriation wrapper functions.}
}
\value{
A tibble with the seriation result in a long format.
Additional values are stored in object attributes. See \code{attributes(result)$raw}.

row: Character. Names of rows.

row_order: Integer. Order of rows.

col: Character. Names of columns.

col_order: Integer. Order of columns.

value: Numeric. Value in input matrix for respective row and col.
}
\description{
Seriation function wrappers that give the result in a long tidy data.frame.
The result table can be transformed to a wide format with \code{seriation2widedf}.
}
\examples{
seriation.seriation_seriate(matuskovo_material, method = "PCA")

matuskovo_CountMatrix <- as(matuskovo_material, "CountMatrix")
seriation.tabula_seriate(matuskovo_CountMatrix, method = "correspondance")

# transform back to wide format
spread_seriation(seriation.seriation_seriate(matuskovo_material, method = "PCA"))

}
