% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriation_utility_functions.R
\name{tidyseriation}
\alias{tidyseriation}
\title{tidyseriation}
\usage{
tidyseriation(x, method = "PCA", control = NULL, ...)
}
\arguments{
\item{x}{test}

\item{method}{test}

\item{control}{test}

\item{...}{test}
}
\description{
tidyseriation
}
