% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriation_utility_functions.R
\name{tidyseriation}
\alias{tidyseriation}
\title{tidyseriation}
\usage{
tidyseriation(x, dist_method = "euclidian",
  seriate_method = "Spectral")
}
\arguments{
\item{x}{test}

\item{dist_method}{test}

\item{seriate_method}{test}
}
\description{
tidyseriation
}
