% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriation_utility_functions.R
\name{tidyseriation}
\alias{tidyseriation}
\title{tidyseriation}
\usage{
tidyseriation(x, dist_args = list(method = "euclidian"),
  seriate_args = list(method = "Spectral"))
}
\arguments{
\item{x}{test}

\item{dist_args}{test}

\item{seriate_args}{test}
}
\description{
tidyseriation
}
